14-LB1
That is the final version of the LB1 repository. Now the Java Code has it's own folder to keep everything clean. To see my code with different API's which I often changed to make user-experienc every time better I recommend looking in the history of the repository.If you have any questions feel free to ask.